Implemented (sort-of) functional color-picker
The messageAppender now sets the receiving Address on the message so that services can reply to only this message.
Added a ColorPicker to the html, allowing you to pick the new color of a device.
ColorPicker values are now sent back to the color-capable device.

Broken in opera, colorPicker is displayed but values not read.
Broken in firefox, colorPicker not available but values are sent to server.