Discovery and lights control are now handled using HazelCast.
The raspberry pi had problems joining the cluster. Communicating over hz is more reliable.

Fixed some memory leaks by Threads not getting destroyed on application shutdown
Sunset mode has been removed from lights control in favor of sensor data
Runmodes are now the responsibility of the lights-api and are no longer part of the Capabilities.
Capabilities are now no longer extendible
JNA is now a provided library and must be placed on the servlet container
Introduced a ClusterBusiness. It will take care of registering itself on the network for auto-discovery.
Refactored all hazelcast specifics (except for Entries, todo). All is hidden in the business components