A Hazelcast server is now required as a separate deployment
When deploying a second webapp on the pi with a hazelcast node, the performance would be so terrible that it would be continuously migrating data between nodes and losing connections. It is now required to perform a separate deployment of the hazel-server before starting the hazel-client nodes.

Added a new dist. package for the hazelcast server
Default setting for any application is to join a hazelcast server as a localhost client